pool: add option to disable packet pool
This patch is adding an option named 'no-packet-pool'. If set to
yes, the packet pool will not be allocated and instead direct
allocation will be used. The default value of the option is 'no'
to keep backward compatibility.

This option has proven to have a huge impact on some system where
the packet pool handling was taking a huge amount of CPU in the
spinlock.