netmap: fix promiscous setting for FreeBSD
This patch set permanent promiscous mode on FreeBSD. This is the
only mode supported and this is quivalent to do something like:
 ifconfig em0 promisc
Reverting state is not considered as there is no way to guarantee
it will occur. Second reason is that the iface has to be dedicated
to netmap (it even loses address when suricata is started).