Debug-validation: track thread id in flow
In workers mode, all packets of a single flow should be handled by
a single thread. In autofp mode, all packets in a flow should be
handled by the same stream/detect thread.

This patch stores the id in the flow and aborts if more than one
thread handling a flow.

The check and store are done from the stream engine, so the autofp
case works as well.