flow: multiple app layer support
This patch adds support for application layer reset. This is
needed for protocol like HTTP and SMTP. For example, in HTTP
after a succesful CONNECT message the application layer linked
with the flow will change to another protocol (usually HTTPS).

To do so, the application layer attached to the Flow has been
transformed to an array to be able to have a list of chained
protocols. For that a structure named FlowAppLayer has been
defined and an array of FlowAppLayer has been attached to
the Flow via the applayer field.

Access and setting of the application layer related elements
in the Flow structure are changed. Accessing elements is
obtained via the FlowGetApp* functions family. Setting
elements is done via FlowSetApp* funtions family.

At this stage, this new feature is unused but it will be used
when implementing the reset of application layer needed by the
operation similar to HTTP CONNECT message.

The modification on flow has been obtained via coccinelle
transformations and a few manual edit. Here's the coccinelle
semantic patches:

First patch:

 @@
 struct Flow_ *f;
 @@

 (
 - f->alparser
 + FLOW_GET_CURRENT_APPLAYER(f).alparser
 |
 - f->alproto
 + FLOW_GET_CURRENT_APPLAYER(f).alproto
 |
 - f->alstate
 + FLOW_GET_CURRENT_APPLAYER(f).alstate
 )

 @@
 Flow *f;
 @@

 (
 - f->alparser
 + FLOW_GET_CURRENT_APPLAYER(f).alparser
 |
 - f->alproto
 + FLOW_GET_CURRENT_APPLAYER(f).alproto
 |
 - f->alstate
 + FLOW_GET_CURRENT_APPLAYER(f).alstate
 )

 @@
 Flow f;
 @@

 (
 - f.alparser
 + FLOW_GET_CURRENT_APPLAYER(&f).alparser
 |
 - f.alproto
 + FLOW_GET_CURRENT_APPLAYER(&f).alproto
 |
 - f.alstate
 + FLOW_GET_CURRENT_APPLAYER(&f).alstate
 )

Second patch:

 @@
 Flow *f;
 expression E;
 @@

 (
 - FLOW_GET_CURRENT_APPLAYER(f).alproto = E
 + FlowSetAppProtocol(f, E)
 |
 - FLOW_GET_CURRENT_APPLAYER(f).alstate = E
 + FlowSetAppState(f, E)
 |
 - FLOW_GET_CURRENT_APPLAYER(f).alparser = E
 + FlowSetAppParser(f, E)
 )

 @@
 Flow *f;
 @@

 (
 - FLOW_GET_CURRENT_APPLAYER(f).alproto
 + FlowGetAppProtocol(f)
 |
 - FLOW_GET_CURRENT_APPLAYER(f).alstate
 + FlowGetAppState(f)
 |
 - FLOW_GET_CURRENT_APPLAYER(f).alparser
 + FlowGetAppParser(f)
 )

 @@
 struct Flow_ *f;
 @@

 (
 - FLOW_GET_CURRENT_APPLAYER(f).alproto
 + FlowGetAppProtocol(f)
 |
 - FLOW_GET_CURRENT_APPLAYER(f).alstate
 + FlowGetAppState(f)
 |
 - FLOW_GET_CURRENT_APPLAYER(f).alparser
 + FlowGetAppParser(f)
 )