TLS: reset hs_bytes_processed when resetting TLS parser state
After every new TLS record, the hs_bytes_processed counter needs to be
reset, or the next record will be parsed with an offset for the second
record.
When analysing the ServerHello message, this causes the
ServerCertificate message to be parsed incorrectly (it was not detected
previously because only the ServerCertificate was analyzed).