output-json-http: add proxy username logging
This patch adds proxy username logging to EVE HTTP event. If Basic
authentication is used on the proxy, suricata extract username
by using the proxy-authorization header.

To activate the feature, you need to add 'proxy_username' in the
custom fields to log:
 - http:
  extended: yes     # enable this for extended logging information
  custom: [proxy_username]

This patch updates the logic of the http field by adding a callback
function that can be used to transform the string we get for libhtp.