detect: avoid race condition on app layer used
Under serious load, it is possible that a app layer get changed on
a flow when another packet of the Flow is still examined in Detect.
The consequence is that it is possible to app layer to get updated
and the rest of the detection run with the new app layer that don't
have the same property (such as Tx handling function which are called
on TLS session).

This is mainly the case when alproto is fetched from Flow and then
the Flow is unlocked and modifiable by another thread. When alstate
is fetch later, we can have alstate not matching alproto and this
causes crashes.

To fix that this patch is using alindex to access to the original
application layer during the detection. This means some function
prototype have been update to use alindex instead of alproto.
Also the FlowGet*AtIndex function are used with alindex param to
access to the correct application layer.

For reference, here's one the backtrace:

 (gdb) bt
 #0  0x0000000000000000 in ?? ()
 #1  0x00000000004310e3 in AppLayerParserSetTransactionInspectId (pstate=0x151743b10, ipproto=ipproto@entry=6 '\006', alproto=alproto@entry=4, alstate=alstate@entry=0x14d47b790, direction=direction@entry=4 '\004')
     at app-layer-parser.c:536
 #2  0x000000000048e6d7 in DeStateUpdateInspectTransactionId (f=0x7ffefc68ba00, direction=4 '\004') at detect-engine-state.c:785
 #3  0x000000000045c034 in SigMatchSignatures (th_v=0x3420fa50, de_ctx=0x23b6f00, det_ctx=0x13fd68ea0, p=<optimized out>) at detect.c:1589
 #4  0x000000000045c9f3 in Detect (data=<optimized out>, p=<optimized out>, tv=<optimized out>, pq=<optimized out>, postpq=<optimized out>) at detect.c:1744
 #5  Detect (tv=<optimized out>, p=<optimized out>, data=<optimized out>, pq=<optimized out>, postpq=<optimized out>) at detect.c:1716
 #6  0x000000000053d24d in TmThreadsSlotVarRun (tv=0x3420fa50, p=0x13fd56920, slot=0x14d47b790, slot@entry=0x13db1ecc0) at tm-threads.c:575
 #7  0x00000000005186ba in TmThreadsSlotProcessPkt (p=0x13fd56920, s=0x13db1ecc0, tv=0x3420fa50) at tm-threads.h:148
 #8  AFPReadFromRing (ptv=ptv@entry=0x13b029bd0) at source-af-packet.c:875
 #9  0x000000000051b5fd in ReceiveAFPLoop (tv=<optimized out>, data=0x13b029bd0, slot=<optimized out>) at source-af-packet.c:1215
 #10 0x00000000005408db in TmThreadsSlotPktAcqLoop (td=0x3420fa50) at tm-threads.c:722
 #11 0x00007ffff6920b50 in start_thread () from /lib/x86_64-linux-gnu/libpthread.so.0
 #12 0x00007ffff51e8a7d in clone () from /lib/x86_64-linux-gnu/libc.so.6
 #13 0x0000000000000000 in ?? ()