pfring: IPS/TAP feature
This patch adds new variables to pfring configuration:

  copy-mode
  copy-iface
  flush-packet

If copy-mode is set to ips, the packet with action DROP
are not copied to the destination interface.

If copy-mode is set to tap, all packets are copied to the
destination interface.

If flush-packet is set to no, you will decrease your CPU usage but
at the cost of sending packets in trains and thus at larger latency.