Removed unneeded code and reworked metadata generating.
After merging the functions from log-file and log-filestore
alot of work was needed to make the code be able to use the
same functions. The code was origionaly copied from log-filestore
with the methods used there to see if json was enabled or not.
Some parts of the code in f.ex LogFileMetaGetUri and similar functions
were then redundent and were deleted. Also workd on how the JSON structure
was built and sent around. Also i removed a return from a void
function (static void LogFileLogDeInitCtx) that was not needed after /
correct after the merger.