yaml: change default layout
Reshuffle yaml file to put values that are most likely to be changed on
top, and values that are advanced and least likely to be changed down.

Add 'step 1' to 'step 5' to indicate to the user that these 5 steps
should be reviewed.

Group cpature methods together, IPS modes together, threading settings
together, etc.