Move tcp.reassembly_memuse to flow manager thread
This solves two problems. One is that every worker thread was reporting
the same value for TCP reassembly memuse, which implied that it was a
per-thread value, but it is global.

The other problem is that the true value, stored in the global ra_memuse, was
only copied to the tcp_reassembly_memuse stat in one function, when it was
increased, so it missed when the value went down.