files: allow hashes for truncated files and add to http log
Files are truncated when they exceed stream reassembly depth. However, for some
purposes, a hash of the prefix of the file is still useful, as
long as we know the size of this prefix. We therefore change the logging logic
to output the hash in the case of file truncation.