Allow md5 hashes for truncated files and add to http log
Files are truncated when they exceed stream reassembly depth. However, for our
purposes (e.g., Nazca), a hash of the prefix of the file is still useful, as
long as we know the size of this prefix. We therefore change the logging logic
to output the hash in the case of file truncation.

We also add the md5 hash and the size of the file (potentially truncated) over
which it was calculated to the HTTP log. These two values are output if the
format specifier %M is used. They are separated by a space. If no hash was
calculated (e.g., nothing was downloaded or md5 hashing is disabled), a - is
output instead.