output-json: reuse OutputJsonCtx in submodules
This patch updates the OutputJsonBuffer function to have it use
on OutputJsonCtx as parameter. This allow to respect the choice
of logging type (syslog, file) done on the eve-json instance.

This fixes the problem of not being able to use two eve-json
instance with different logging method.

An update of all per-protocol structures has been necessary to do
that.