files: file-hash-byte-limit option
Allow capping number of bytes of files used for file hash. This is to allow
for consistency and avoid relying on whatever the stream reassembly limit
happens to be. If 0 or not present, the original behaviour of hashing up to the
stream reassembly limit remains. If the value is greater than the stream
reassembly limit, the stream reassembly limit takes precedence.