unix socket: protocol v0.2
This patch updates the unix socket protocol. Messages send from
the server have now a '\n' at the end. This allows the client to
detect easily the end of a command.

Backward compatility is preserved as a client with the older
version of the protocol can still connect to a Suricata with
version 2 of the protocol.