ConfGet - fail if the node value is 0.
Found while running AFL - many users of ConfGet will assume that
value is usable if the function returns success, but this function
will return success if the node exists, but has a NULL value.

Instead, now fail of the value is NULL.  Fixes a bunch of cases
where users of this function try to do something like strcmp
on a successful result.

A quick review of all users of this function didn't reveal any
cases where this should break things.