util-logopenfile: suppress unused batch timeout
The batch-timeout variable was unused in the code so we can
suppress it for now. It was meant to be able to log in the case of
low trafic suricata by flushing after a timeout. This patch replaces
the feature by a comment in the YAML stating that the pipeling is
meant for high bandwidth suricata.